Update to Rust 1.97
Bump the pinned toolchain and MSRV from 1.89 to 1.97 (latest stable) and
fix the new clippy lints: unnecessary_cast, useless_borrows_in_formatting,
and manual_is_multiple_of.

+++ src/fuse_adapter.rs
@@ -883,7 +883,7 @@ impl Filesystem for FleetFUSE {
                     attr.perm,
                     req.uid(),
                     req.gid(),
-                    mask.bits() as i32,
+                    mask.bits(),
                 ) {
                     reply.ok();
                 } else {
